Squashed 'Externals/Core/' changes from e5f49e3..3d9518d
3d9518d decimal
9ef7d6f GenTypeOfTransactionEnum

git-subtree-dir: Externals/Core
git-subtree-split: 3d9518dd800b22cb044964d78813e33f6329fd32

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/MerchantOfferDto.cs
@@ -27,9 +27,9 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
             }
         }
 
-        public double Credit { get; set; }
+        public decimal Credit { get; set; }
 
-        public double Debit { get; set; }
+        public decimal Debit { get; set; }
 
         public OppContractLengthEnum Duration { get; set; }
